\name{DSi}
\alias{DSi}
\title{Moment Estimators associated with subject-specific latent variable}
\usage{
  DSi(Beta, Zi, X)
}
\value{
  numeric vector
}
\description{
  \eqn{\frac{d}{d\beta} S_i = S_i X_i^T}. The \code{Zi} and
  \code{X} are subject-specific latent varialbe and
  time-independent covariate associated with ordered
  \code{y}
}
\seealso{
  \code{\link{Si}}
}

