\name{F.hat}
\alias{F.hat}
\title{The non-parametric maximum likelihood estimator of shape function.}
\usage{
  F.hat(obj)
}
\arguments{
  \item{obj,}{an instance of class
  \code{\link{recurrent-data}}}
}
\value{
  The estimator based on the given recurrent-data. The
  estimator is a function \code{f(t)}.
}
\description{
  This function implemented the equation (4) in Mei-Cheng
  Wang, Jing Qin, and Chin-Tsang Chiang 2004.
}
\references{
  Mei-Cheng Wang, Jing Qin, and Chin-Tsang Chiang 2004.
}

