\name{Lambda.hat}
\alias{Lambda.hat}
\title{non-parametric maximum likelihood estimator of cumulative rate function(CRF).}
\usage{
  Lambda.hat(obj, bootstrap = FALSE)
}
\arguments{
  \item{obj,}{an instance of class
  \code{\link{recurrent-data}}}

  \item{obj,}{an instance of class
  \code{\link{recurrent-data}}}
}
\description{
  non-parametric maximum likelihood estimator of cumulative
  rate function(CRF).
}

