\name{recurrent-data}
\alias{recurrent-data}
\title{Recurrent Data}
\arguments{
  \item{X,}{data.frame. The time independent covariate of
  subjects.}

  \item{y,}{numeric vector. The censor time.}

  \item{t,}{list. The time of recurrent events.}

  \item{W,}{data.frame. The time dependent covariates.}

  \item{T_0,}{numeric value. The time of termination of
  experients.}
}
\description{
  The S4 class for recurrent data.
}

