\name{U.hat.gen}
\alias{U.hat.gen}
\title{Formulat(3) in Huang2004}
\usage{
  U.hat.gen(obj, Zi = NULL, ...)
}
\arguments{
  \item{obj}{recurrent-data object}

  \item{Zi}{numeric vector, \eqn{\hat{Z}_i}}

  \item{...}{arguments pass to \code{\link{Zi.hat}}}
}
\value{
  numeric vector with length \code{ncol(obj
}
\description{
  Formulat(3) in Huang2004
}

